data_lake/data_broker/container/requirements/base.txt
azure-identity==1.12.0
azure-keyvault-secrets==4.6.0
azure-storage-blob==12.14.1
django-ipware==4.0.2
Django==4.1.6
djangorestframework==3.14.0
psycopg2-binary==2.9.5
